changed report exporters to new jasperreports API exporters, code clean up
git-svn-id: http://svn.code.sf.net/p/dynamicreports/code/trunk@815 bf0c503b-2351-4f66-9eae-aa52f3279a3f

+++ dynamicreports-core/src/main/java/net/sf/dynamicreports/design/transformation/StyleResolver.java
@@ -93,7 +93,7 @@ public class StyleResolver {
 		}
 
 		FontUtil fontUtil = FontUtil.getInstance(DefaultJasperReportsContext.getInstance());
-		Font font = fontUtil.getAwtFontFromBundles(fontName, fontStyle, fontSize, Locale.getDefault(), true);
+		Font font = fontUtil.getAwtFontFromBundles(fontName, fontStyle, (float) fontSize, Locale.getDefault(), true);
 		if (font == null) {
 			font = new Font(fontName, fontStyle, fontSize);
 		}
